requirements.txt
pywebview==4.4.1
python-dotenv==1.0.1
